added for each check an own file to increade structure and understandability
